Do not create personal collections for API keys
Functions like `user->personal-collection-id` automatically create a
personal collection. These were called during permissions checks and
some hydration functions for the User API.

One complicating factor here is that `:type` is not typically one of the
columns returned when you do `(t2/select :model/User)`. This means that
even when we have a `user` available, we usually can't cheaply check to
see whether it's an API key user (this might be something we should
change, but I think it would be more involved and still wouldn't
guarantee that the `type` is present when we need it).

It would also be potentially expensive to maintain a guarantee that we
never insert a personal collection for an API key user, because we'd
need to do that in the `before-insert` hook for collections, grabbing
the relevant `:model/User`. Of course, we only need to do this for
collections with a `personal_owner_id`, and maybe in production not many
collections are going to be created at once, but...

The second point here is that if we do end up creating a personal
collection for an API key user, it's not a huge problem - we definitely
shouldn't do it, but I think in production it's probably better to
create it rather than throwing an exception.

As a compromise, I added a check that asserts that the user isn't an API
key user, but only in dev/test environments, not prod. This way, if we
added a codepath that created one, we'd hopefully catch it there.